requirements.txt
Django==3.1.2
django-admin-sortable2==0.7.7
django-tinymce==3.1.0
environs==8.0.0
Pillow==8.0.0
python-dotenv==0.9.1
requests==2.24.0
